completing SNA-14
