+++ src/screens/EnterScreen/types.ts
@@ -1,0 +1,7 @@
+import { NativeStackScreenProps } from '@react-navigation/native-stack'
+import { RootStackParamList } from '../../App'
+
+export type EnterScreenProps = NativeStackScreenProps<
+	RootStackParamList,
+	'Enter'
+>
